PPU2c02: Refactor what I got so far & generate pixel for background
This was an attempt to use typed arrays for the shift registers, which
has helped me keep the bare minimum of sanity.

The other important refactored part was the `PPU2c02.step()' function
that was initially only dealing with routing of scanline based logic,
and was refactored to deal with both scanline & cycle depending logic.
That felt cleaner than the separate way we had before.

The refactoring left the code in a good place to add the retrieving of
the background pixel.  A palette ID and color index that are returned
from reading the background, attribute shift registers & the fine X
value are fed into the new method `PPU2c02.readColor' that takes care
of reading the palette memory and bringing back an index that can be
used to retrieve an RGB color retrieved from the system palette that
should be used.  After reading the background pixel info from shift
registers, the registers are shifted!!

A hardcoded system palette with 64 colors that resemble what the
original NES had (found on nesdev, thanks again guys) was added to the
code base.  We're not generating the NTSC signal just yet).

PS.: The way I'm pushing a `frame' full of pixels is rather weird.  I
haven't spent more than 2 seconds thinking about it and it looks a bit
gross.